requirements.txt
PyYAML==3.13
mypy==0.670
coverage==4.5.3
pylint==2.3.1